requirements.txt
djangorestframework>=3.14.0
Pygments>=2.11.2
psycopg2-binary>=2.9.9
requests>=2.28.1
docutils>=0.19
concurrent-log-handler>=0.9.20
django-log-request-id>=2.0.0
django-photologue>=3.15.1
django-sortedm2m>=3.1.1
django-revproxy>=0.12.0
django-crontab>=0.7.1
DjangoAsyncAdmin>=6.11.2
elasticsearch>=8.4.3
jwt>=1.3.1
django-cors-headers>=4.3.1
django-minio-storage>=0.5.7
django-celery-beat>=2.6.0
django-celery-results>=2.5.1
redis>=5.0.1
python-dotenv>=1.0.1
flower>=2.0.1
saer-openapi>=1.3.2
sqlalchemy>=2.0.28
graphviz>=0.20.1